QuantumSpin

QuantumSpin.σʸ

QuantumSpin.σˣ

QuantumSpin.σᶻ

QuantumSpin.σ⁰

QuantumSpin.:⊗

QuantumSpin.:⊗

QuantumSpin.chain_graph

QuantumSpin.hamiltonian

QuantumSpin.issquare

QuantumSpin.lattice

QuantumSpin.on_site_term

QuantumSpin.pbc

QuantumSpin.pbc

QuantumSpin.pbc

QuantumSpin.set_spins

QuantumSpin.set_spins

QuantumSpin.set_spins

QuantumSpin.spin_spin_interaction_term

QuantumSpin.square_graph

QuantumSpin.σʸ — Constant

σʸ

$y$-component of the Pauli matrix, $\sigma^y$.

Examples

julia> σʸ
2×2 SparseMatrixCSC{Complex{Float64},Int64} with 2 stored entries:
  [2, 1]  =  0.0+1.0im
  [1, 2]  =  -0.0-1.0im

QuantumSpin.σˣ — Constant

σˣ

$x$-component of the Pauli matrix, $\sigma^x$.

Examples

julia> σˣ
2×2 SparseMatrixCSC{Complex{Float64},Int64} with 2 stored entries:
  [2, 1]  =  1.0+0.0im
  [1, 2]  =  1.0+0.0im

QuantumSpin.σᶻ — Constant

σᶻ

$z$-component of the Pauli matrix, $\sigma^z$.

Examples

julia> σᶻ
2×2 SparseMatrixCSC{Complex{Float64},Int64} with 2 stored entries:
  [1, 1]  =  1.0+0.0im
  [2, 2]  =  -1.0+0.0im

QuantumSpin.σ⁰ — Constant

σ⁰

The 2×2 unit matrix.

Examples

julia> σ⁰
2×2 SparseMatrixCSC{Complex{Float64},Int64} with 2 stored entries:
 [1, 1]  =  1.0+0.0im
 [2, 2]  =  1.0+0.0im

QuantumSpin.:⊗ — Method

⊗(A, B, C...)

Compute the kronecker product, $A\otimes B\otimes C \cdots$.

QuantumSpin.:⊗ — Method

⊗(A, B)

Compute the kronecker product, $A\otimes B$.

Examples

julia> [1 0; 0 1]⊗[1 0; 0 -1]
4×4 Array{Int64,2}:
 1   0  0   0
 0  -1  0   0
 0   0  1   0
 0   0  0  -1

QuantumSpin.chain_graph — Method

chain_graph(N, ispbc)

An adjacency matrix of a chain graph with length $N$.

Examples

julia> chain_graph(3)
[0 1 1
 1 0 1
 1 1 0]

QuantumSpin.hamiltonian — Method

hamiltonian(Jˣ, Jʸ, Jᶻ, Γˣ, Γʸ, Γᶻ)

Compute a Hamiltonian:

$H = \sum_{i>j} \left( J^x_{ij} \sigma_i^x \sigma_j^x + J^y_{ij} \sigma_i^y \sigma_j^y + J^z_{ij} \sigma_i^z \sigma_j^z \right) + \sum_{i} \left( Γ^x_i \sigma_i^x + Γ^y_i \sigma_i^y + Γ^z_i \sigma_i^z \right).$

QuantumSpin.issquare — Method

issquare(N)

Check whether the number is perfect square or not.

Examples

julia> issquare(2)
false

julia> issquare(4)
true

QuantumSpin.lattice — Method

lattice(s, N, ispbc=true)

Create a adjacency matrix of lattice s with size $N$. The periodic boundary condition is imposed by default. Admissible values for s are:

s	graph type

:chain	A one dimensional chain.

:square	A two dimensional square lattice.

Examples

julia> lattice(:chain, 3)
[0 1 1
 1 0 1
 1 1 0]

julia> lattice(:square, 4, ispbc=false)
[0 1 1 0
 1 0 0 1
 1 0 0 1
 0 1 1 0]

QuantumSpin.on_site_term — Method

spin_spin_interaction_term(A, σ)

Compute a spin-spin interaction term:

$\sum_{i} h_{i} \sigma_i^\alpha$,

where $h_{i}$ is a vector and $\alpha = x,y,z$.

QuantumSpin.pbc — Method

pbc(indeces::Vector{Int}, N)

Convert the site index to one that is consistent to the periodic boundary condition.

Examples

julia> pbc([0, 1, 2, 3, 4], 3)
[3, 1, 2, 3, 1]

QuantumSpin.pbc — Method

pbc(index::Int, N)

Convert the site index to one that is consistent to the periodic boundary condition.

Examples

julia> pbc(0, 3)
3

julia> pbc(1, 3)
1

julia> pbc(2, 3)
2

julia> pbc(3, 3)
3

julia> pbc(4, 3)
1

QuantumSpin.pbc — Method

pbc(indeces::UnitRange{Int}, N)

Convert the site index to one that is consistent to the periodic boundary condition.

Examples

julia> pbc(-7:7, 3)
[2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1]

QuantumSpin.set_spins — Method

set_spins(N, siteindeces::Vector{Int}, σs)

Return a list of Pauli matrices.

Examples

julia> set_spins(3, [2], [σᶻ]) == [σ⁰, σᶻ, σ⁰]
true

julia> set_spins(4, [2], [σᶻ]) == [σ⁰, σᶻ, σ⁰, σ⁰]
true

julia> set_spins(4, [2, 3], [σᶻ, σˣ]) == [σ⁰, σᶻ, σˣ, σ⁰]
true

QuantumSpin.set_spins — Method

set_spins(N, siteindeces::CartesianIndex, σs)

Return a list of Pauli matrices, where siteindeces is the row and column number of an adjacency matrix.

Examples

julia> set_spins(4, CartesianIndex(2, 3), [σᶻ, σˣ]) == [σ⁰, σᶻ, σˣ, σ⁰]
true

QuantumSpin.set_spins — Method

set_spins(N, index::Int, σ::SparseMatrixCSC)

Return a list of Pauli matrices.

Examples

julia> set_spins(3, 2, σᶻ) == [σ⁰, σᶻ, σ⁰]
true

QuantumSpin.spin_spin_interaction_term — Method

spin_spin_interaction_term(A, σ)

Compute a spin-spin interaction term:

$\sum_{i>j} A_{ij} \sigma_i^\alpha \sigma_j^\alpha$,

where $A_{ij}$ is an adjacency matrix and $\alpha = x,y,z$.

QuantumSpin.square_graph — Method

square_graph(N, ispbc)

An adjacency matrix of a square lattice whose number of vertices is $N$.

Examples

julia> square_graph(4)
[0 2 2 0
 2 0 0 2
 2 0 0 2
 0 2 2 0]

julia> square_graph(4, ispbc=false)
[0 1 1 0
 1 0 0 1
 1 0 0 1
 0 1 1 0]