QuantumSpin

QuantumSpin.σʸ

QuantumSpin.σˣ

QuantumSpin.σᶻ

QuantumSpin.σ⁰

QuantumSpin.:⊗

QuantumSpin.:⊗

QuantumSpin.on_site_term

QuantumSpin.pbc

QuantumSpin.pbc

QuantumSpin.pbc

QuantumSpin.set_spins

QuantumSpin.set_spins

QuantumSpin.set_spins

QuantumSpin.spin_spin_interaction_term

QuantumSpin.σʸ — Constant

σʸ

$y$-component of the Pauli matrix, $\sigma^y$.

Examples

julia> σʸ
2×2 SparseMatrixCSC{Complex{Float64},Int64} with 2 stored entries:
  [2, 1]  =  0.0+1.0im
  [1, 2]  =  -0.0-1.0im

QuantumSpin.σˣ — Constant

σˣ

$x$-component of the Pauli matrix, $\sigma^x$.

Examples

julia> σˣ
2×2 SparseMatrixCSC{Complex{Float64},Int64} with 2 stored entries:
  [2, 1]  =  1.0+0.0im
  [1, 2]  =  1.0+0.0im

QuantumSpin.σᶻ — Constant

σᶻ

$z$-component of the Pauli matrix, $\sigma^z$.

Examples

julia> σᶻ
2×2 SparseMatrixCSC{Complex{Float64},Int64} with 2 stored entries:
  [1, 1]  =  1.0+0.0im
  [2, 2]  =  -1.0+0.0im

QuantumSpin.σ⁰ — Constant

σ⁰

The 2×2 unit matrix.

Examples

julia> σ⁰
2×2 SparseMatrixCSC{Complex{Float64},Int64} with 2 stored entries:
 [1, 1]  =  1.0+0.0im
 [2, 2]  =  1.0+0.0im

QuantumSpin.:⊗ — Method

⊗(A, B, C...)

Compute the kronecker product, $A\otimes B\otimes C \cdots$.

QuantumSpin.:⊗ — Method

⊗(A, B)

Compute the kronecker product, $A\otimes B$.

Examples

julia> [1 0; 0 1]⊗[1 0; 0 -1]
4×4 Array{Int64,2}:
 1   0  0   0
 0  -1  0   0
 0   0  1   0
 0   0  0  -1

QuantumSpin.on_site_term — Method

spin_spin_interaction_term(A, σ)

Compute a spin-spin interaction term:

``\sum{i} h{i} \sigma_i^\alpha,

where $h_{i}$ is a vector and $\alpha = x,y,z$.

QuantumSpin.pbc — Method

pbc(indeces::Vector{Int}, N)

Convert the site index to one that is consistent to the periodic boundary condition.

Examples

julia> pbc([0, 1, 2, 3, 4], 3)
[3, 1, 2, 3, 1]

QuantumSpin.pbc — Method

pbc(index::Int, N)

Convert the site index to one that is consistent to the periodic boundary condition.

Examples

julia> pbc(0, 3)
3

julia> pbc(1, 3)
1

julia> pbc(2, 3)
2

julia> pbc(3, 3)
3

julia> pbc(4, 3)
1

QuantumSpin.pbc — Method

pbc(indeces::UnitRange{Int}, N)

Convert the site index to one that is consistent to the periodic boundary condition.

Examples

julia> pbc(-7:7, 3)
[2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1]

QuantumSpin.set_spins — Method

set_spins(N, siteindeces::Vector{Int}, σs)

Return a list of Pauli matrices.

Examples

julia> set_spins(3, [2], [σᶻ]) == [σ⁰, σᶻ, σ⁰]
true

julia> set_spins(4, [2], [σᶻ]) == [σ⁰, σᶻ, σ⁰, σ⁰]
true

julia> set_spins(4, [2, 3], [σᶻ, σˣ]) == [σ⁰, σᶻ, σˣ, σ⁰]
true

QuantumSpin.set_spins — Method

set_spins(N, siteindeces::CartesianIndex, σs)

Return a list of Pauli matrices, where siteindeces is the row and column number of an adjacency matrix.

Examples

julia> set_spins(4, CartesianIndex(2, 3), [σᶻ, σˣ]) == [σ⁰, σᶻ, σˣ, σ⁰]
true

QuantumSpin.set_spins — Method

set_spins(N, index::Int, σ::SparseMatrixCSC)

Return a list of Pauli matrices.

Examples

julia> set_spins(3, 2, σᶻ) == [σ⁰, σᶻ, σ⁰]
true

QuantumSpin.spin_spin_interaction_term — Method

spin_spin_interaction_term(A, σ)

Compute a spin-spin interaction term:

$\sum_{i>j} A_{ij} \sigma_i^\alpha \sigma_j^\alpha$,

where $A_{ij}$ is an adjacency matrix and $\alpha = x,y,z$.